Refactor fetchPredefinedProperties to use fetchStudyMetadata instead of fetchAppsAndUrls. Add fetchDefaultCountry function to set default country in SubstationCreationForm. Add metadata.ts to fetch study metadata.
