+++ src/components/dialogs/network-modifications/substation/creation/substation-creation-form.jsx
@@ -7,7 +7,7 @@
 
 import Grid from '@mui/material/Grid';
 import { filledTextField, gridItem } from '../../../dialogUtils';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { TextInput } from '@gridsuite/commons-ui';
 import {
     COUNTRY,
@@ -16,8 +16,21 @@ import {
 } from 'components/utils/field-constants';
 import CountrySelectionInput from 'components/utils/rhf-inputs/country-selection-input';
 import PropertiesForm from '../../common/properties/properties-form';
+import { fetchDefaultCountry } from 'components/dialogs/network-modifications/common/properties/property-utils';
+import { useFormContext } from 'react-hook-form';
 
 const SubstationCreationForm = () => {
+    const { setValue } = useFormContext();
+
+    // We set the default country if there is one
+    useEffect(() => {
+        fetchDefaultCountry().then((country) => {
+            if (country) {
+                setValue(COUNTRY, country);
+            }
+        });
+    }, [setValue]);
+
     const substationIdField = (
         <TextInput
             name={EQUIPMENT_ID}
